add separate css file, handle outside click event
