+++ src/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
-import logo from './logo.svg';
 import './App.css';
-import Modal from './modal';
+import Modal from './modal/Modal';
 
 class App extends React.Component{
   state = {
